feat: UserPerformance 삭제 API 구현

+++ src/main/java/bandwagon/bandwagonback/service/UserPerformanceService.java
@@ -37,7 +37,18 @@ public class UserPerformanceService {
 
     // 기존 연주기록 삭제
     @Transactional
-    public void deleteUserPerformance(Long userPerformanceId){
+    public void deleteUserPerformance(String email, Long userPerformanceId) throws Exception {
+        User user = userRepository.findByEmail(email).orElse(null);
+        if (user == null) {
+            throw new Exception("User does not exist!");
+        }
+        UserPerformance userPerformance = userPerformanceRepository.findById(userPerformanceId).orElse(null);
+        if (userPerformance == null) {
+            throw new Exception("User Performance does not exist!");
+        }
+        if (userPerformance.getUser() != user) {
+            throw new Exception("Owner of performance is not inputted user!");
+        }
         userPerformanceRepository.deleteById(userPerformanceId);
     }
